requirements.txt
faiss-cpu==1.9.0
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
openai==1.44.0
pypdf==5.1.0
PyPDF2==3.0.1
python-dotenv==1.0.1
sentence-transformers==3.2.1
streamlit==1.38.0
streamlit-chat==0.1.1
tiktoken==0.7.0